requirements.txt
# BioFace — Davomat Tizimi
# To'liq bog'liqliklar ro'yxati
#
# O'rnatish: pip install -r requirements.txt

# ── Web Framework ──────────────────────
fastapi==0.135.1
uvicorn[standard]==0.41.0

# ── Database ───────────────────────────
SQLAlchemy==2.0.46

# ── Templates ──────────────────────────
Jinja2==3.1.6
itsdangerous==2.2.0

# ── Form / File Upload ──────────────────
python-multipart==0.0.22
Pillow==12.1.1

# ── Telegram Bot ────────────────────────
python-telegram-bot>=21.0,<23.0

# ── Timezone Data ───────────────────────
tzdata>=2025.1

# ── Optional: SMS Gateway (kelajakda) ───
# requests==2.31.0

# ── HTTP Client (ISUP server proxy) ─────
httpx==0.28.1
bcrypt==5.0.0
redis==7.3.0
psutil==7.2.2
fpdf2==2.8.3
hsemotion==0.3.0
